Added some basic Maneuvering code. Doesn't work quite right yet.

+++ src/com/nsxwing/gamestate/Coordinate.groovy
@@ -4,6 +4,6 @@ package com.nsxwing.gamestate
  * Simple class used to mark x and y coordinates of a point.
  */
 public class Coordinate {
-    int x
-    int y
+    double x
+    double y
 }
